tests/wpt/web-platform-tests/fullscreen/api/document-onfullscreenerror.html

<!DOCTYPE html>
<title>Document.onfullscreenerror</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function(t)
{
    var e = document.createElement('span');
    var sync = true;
    document.addEventListener('fullscreenerror', function() {
        assert_false(sync);
    });
    e.requestFullscreen();
    sync = false;
});
</script>
